store-backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
idna==3.10
PyJWT==2.8.0
pytz==2025.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
